Update PromQL till March 30

This commit updates the PromQL code upto commit
3e4bd4d9135200f6e74a2ba3ef47dba1de8656d9

This commit also includes the changes for the new
PromQL stats support which is an optional feature.

+++ pkg/promql/value.go
@@ -166,10 +166,10 @@ func (m Matrix) String() string {
 }
 
 // TotalSamples returns the total number of samples in the series within a matrix.
-func (m Matrix) TotalSamples() int64 {
-	var numSamples int64
+func (m Matrix) TotalSamples() int {
+	var numSamples int
 	for _, series := range m {
-		numSamples += int64(len(series.Points))
+		numSamples += len(series.Points)
 	}
 	return numSamples
 }
